+++ estyp/__init__.py
@@ -1,2 +1,1 @@
-__version__ = "0.9.0"
-print("Looking for help? Visit the ESTYP documentation: https://estyp.readthedocs.io/")
+__version__ = "0.9.2"